logging_system_project/local_test/requirements.txt
requests>=2.28.0
mysql-connector-python>=8.0.0
pandas>=1.5.0
matplotlib>=3.6.0
jinja2>=3.1.0
colorama>=0.4.0


